hw4 part 2 working

+++ trap.c
@@ -32,6 +32,9 @@ idtinit(void)
   lidt(idt, sizeof(idt));
 }
 
+// added because we removed static from vm.c:mappages() definition.
+int mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm);
+
 //PAGEBREAK: 41
 void
 trap(struct trapframe *tf)
@@ -86,12 +89,39 @@ trap(struct trapframe *tf)
               tf->trapno, cpunum(), tf->eip, rcr2());
       panic("trap");
     }
-    // In user space, assume process misbehaved.
-    cprintf("pid %d %s: trap %d err %d on cpu %d "
-            "eip 0x%x addr 0x%x--kill proc\n",
-            proc->pid, proc->name, tf->trapno, tf->err, cpunum(), tf->eip,
-            rcr2());
-    proc->killed = 1;
+
+    // attempt to alloc the page
+    uint faulting_addr = rcr2();
+    uint pg_start_boundary = PGROUNDDOWN(faulting_addr);
+    // get some physical memory, in the form of a virtual addr in
+    // the kernel space of virutal addresses.
+    char* mem = kalloc();
+    if(mem == 0){
+      cprintf("kalloc() dynamic allocation failed!\n");
+      // deallocuvm(pgdir, newsz, oldsz);
+      // In user space, assume process misbehaved.
+      cprintf("pid %d %s: trap %d err %d on cpu %d "
+              "eip 0x%x addr 0x%x--kill proc\n",
+              proc->pid, proc->name, tf->trapno, tf->err, cpunum(), tf->eip,
+              rcr2());
+      proc->killed = 1;
+      return; // return code in eax
+    }
+    memset(mem, 0, PGSIZE);
+    // now map V2P(mem) starting at the pg_start_boundary (VA) with some permissions
+    if(mappages(proc->pgdir, (char*)pg_start_boundary, PGSIZE, V2P(mem), PTE_W|PTE_U) < 0){
+      cprintf("allocuvm out of memory (2)\n");
+      // deallocuvm(pgdir, newsz, oldsz);
+      kfree(mem);
+
+      // In user space, assume process misbehaved.
+      cprintf("pid %d %s: trap %d err %d on cpu %d "
+              "eip 0x%x addr 0x%x--kill proc\n",
+              proc->pid, proc->name, tf->trapno, tf->err, cpunum(), tf->eip,
+              rcr2());
+      proc->killed = 1;
+      return;
+    }
   }
 
   // Force process exit if it has been killed and is in user space.
